requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
decorator==4.3.0
Django==1.11.13
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
future==0.16.0
geocoder==1.38.1
gspread==3.0.0
httplib2==0.11.3
idna==2.6
Markdown==2.6.11
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2018.4
ratelim==0.1.6
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
